refactor(balance): fold snapshots via SnapshotFold accumulator (#812)
* refactor(balance): move folding to Snapshots

Move new_snapshots/ec_set_snapshots from Balances to snapshot.rs as
Snapshots::from_entries and Snapshots::from_ec_entries. This removes
the Self:: smell and places snapshot construction with the other
helpers.

Extract the version-chaining logic into a SnapshotFold accumulator so
it lives in one place instead of being copy-pasted across both entry
points. The two functions now differ only in whether they fold the
leaf account, stated inline as the domain rule it is. Restore the
tracing spans dropped in the initial move.

Add unit tests for EC set snapshot folding:
- updating an existing EC set balance
- fanning one member into multiple EC ancestor sets

* refactor(balance): effective snapshots onto data

new_effective_snapshots was a stateless construction helper hanging
off the EffectiveBalances orchestrator via Self::. Move it onto
EffectiveBalanceData as into_snapshots(journal_id), which yields
EffectiveBalanceSnapshot directly instead of a 5-tuple that the
caller immediately reassembled.

This drops the intermediate (AccountId, Currency, NaiveDate,
BalanceSnapshot, u32) clump, colocates construction with the data it
is built from, and leaves EffectiveBalances doing only orchestration.

* refactor(balance): assemble ranges on BalanceRange

balance_ranges_from_snapshots was a stateless transform on the
EffectiveBalances orchestrator, building BalanceRange values from
(open, close) bounds. Move that construction onto BalanceRange as
from_bounds and inline the map collection at the caller.

Leaves EffectiveBalances::find_all_in_range doing only orchestration;
the repo-side helper of the same name is unrelated and untouched.
