+++ cala-ledger/src/balance/mod.rs
@@ -233,7 +233,7 @@ impl Balances {
             .repo
             .find_for_update(op, journal.id, &all_involved_balances)
             .await?;
-        let new_balances = Self::new_snapshots(
+        let new_balances = Snapshots::from_entries(
             created_at,
             current_balances,
             &entries,
@@ -395,7 +395,7 @@ impl Balances {
         // transactions exactly as they would with per-transaction reads.
         let mut all_new = Vec::new();
         for tx in group.iter() {
-            let new_balances = Self::ec_set_snapshots(
+            let new_balances = Snapshots::from_ec_entries(
                 tx.created_at,
                 current_balances.clone(),
                 &tx.entries,
@@ -448,400 +448,5 @@ impl Balances {
         }
 
         Ok(())
-    }
-
-    #[instrument(level = "debug", name = "cala_ledger.balances.new_snapshots", skip_all)]
-    fn new_snapshots(
-        time: DateTime<Utc>,
-        mut current_balances: HashMap<(AccountId, Currency), Option<BalanceSnapshot>>,
-        entries: &[EntryValues],
-        mappings: &HashMap<AccountId, Vec<AccountSetId>>,
-    ) -> Vec<BalanceSnapshot> {
-        let mut latest_balances: HashMap<(AccountId, &Currency), BalanceSnapshot> = HashMap::new();
-        let mut new_balances = Vec::new();
-        let empty = Vec::new();
-        for entry in entries.iter() {
-            for account_id in mappings
-                .get(&entry.account_id)
-                .unwrap_or(&empty)
-                .iter()
-                .map(AccountId::from)
-                .chain(std::iter::once(entry.account_id))
-            {
-                let latest =
-                    if let Some(latest) = latest_balances.remove(&(account_id, &entry.currency)) {
-                        new_balances.push(latest.clone());
-                        Some(latest)
-                    } else {
-                        None
-                    };
-                let current = current_balances.remove(&(account_id, entry.currency));
-                let Some(balance) = latest.map(Some).or(current) else {
-                    continue;
-                };
-
-                let new_snapshot = match balance {
-                    Some(balance) => Snapshots::update_snapshot(time, balance, entry),
-                    None => Snapshots::new_snapshot(time, account_id, entry),
-                };
-
-                latest_balances.insert((account_id, &entry.currency), new_snapshot);
-            }
-        }
-        new_balances.extend(latest_balances.into_values());
-        new_balances
-    }
-
-    /// Like [`Self::new_snapshots`] but fans each entry **only** into its
-    /// EC ancestor sets (never the leaf account itself, which the inline
-    /// poster path already maintains). Chains repeated writes to the same
-    /// set within the batch so versions increment correctly.
-    #[instrument(name = "cala_ledger.balances.ec_set_snapshots", skip_all)]
-    fn ec_set_snapshots(
-        time: DateTime<Utc>,
-        mut current_balances: HashMap<(AccountId, Currency), Option<BalanceSnapshot>>,
-        entries: &[EntryValues],
-        ec_mappings: &HashMap<AccountId, Vec<AccountSetId>>,
-    ) -> Vec<BalanceSnapshot> {
-        let mut latest_balances: HashMap<(AccountId, &Currency), BalanceSnapshot> = HashMap::new();
-        let mut new_balances = Vec::new();
-        let empty = Vec::new();
-        for entry in entries.iter() {
-            for account_id in ec_mappings
-                .get(&entry.account_id)
-                .unwrap_or(&empty)
-                .iter()
-                .map(AccountId::from)
-            {
-                let latest =
-                    if let Some(latest) = latest_balances.remove(&(account_id, &entry.currency)) {
-                        new_balances.push(latest.clone());
-                        Some(latest)
-                    } else {
-                        None
-                    };
-                let current = current_balances.remove(&(account_id, entry.currency));
-                let Some(balance) = latest.map(Some).or(current) else {
-                    continue;
-                };
-
-                let new_snapshot = match balance {
-                    Some(balance) => Snapshots::update_snapshot(time, balance, entry),
-                    None => Snapshots::new_snapshot(time, account_id, entry),
-                };
-
-                latest_balances.insert((account_id, &entry.currency), new_snapshot);
-            }
-        }
-        new_balances.extend(latest_balances.into_values());
-        new_balances
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    mod new_snapshots {
-        use super::*;
-
-        use chrono::Utc;
-        use rust_decimal::Decimal;
-        use std::collections::HashMap;
-
-        use cala_types::{
-            balance::BalanceAmount,
-            entry::EntryValues,
-            primitives::{DebitOrCredit, Layer},
-        };
-
-        use crate::primitives::{Currency, EntryId, JournalId, TransactionId};
-
-        fn create_test_entry(
-            units: Decimal,
-            direction: DebitOrCredit,
-            layer: Layer,
-            currency: &str,
-            account_id: AccountId,
-        ) -> EntryValues {
-            EntryValues {
-                id: EntryId::new(),
-                version: 1,
-                transaction_id: TransactionId::new(),
-                journal_id: JournalId::new(),
-                account_id,
-                entry_type: "TEST_ENTRY".to_string(),
-                sequence: 1,
-                layer,
-                currency: currency.parse().unwrap(),
-                direction,
-                units,
-                description: None,
-                metadata: None,
-            }
-        }
-
-        fn create_test_balance_snapshot(
-            account_id: AccountId,
-            journal_id: JournalId,
-            currency: Currency,
-            version: u32,
-        ) -> BalanceSnapshot {
-            let time = Utc::now();
-            let entry_id = EntryId::new();
-            BalanceSnapshot {
-                journal_id,
-                account_id,
-                entry_id,
-                currency,
-                settled: BalanceAmount {
-                    dr_balance: Decimal::ZERO,
-                    cr_balance: Decimal::ZERO,
-                    entry_id,
-                    modified_at: time,
-                },
-                pending: BalanceAmount {
-                    dr_balance: Decimal::ZERO,
-                    cr_balance: Decimal::ZERO,
-                    entry_id,
-                    modified_at: time,
-                },
-                encumbrance: BalanceAmount {
-                    dr_balance: Decimal::ZERO,
-                    cr_balance: Decimal::ZERO,
-                    entry_id,
-                    modified_at: time,
-                },
-                version,
-                modified_at: time,
-                created_at: time,
-            }
-        }
-
-        #[test]
-        fn new_snapshots_creates_new_snapshot_when_no_current_balance() {
-            let account_id = AccountId::new();
-            let currency: Currency = "USD".parse().unwrap();
-
-            let entry = create_test_entry(
-                Decimal::from(100),
-                DebitOrCredit::Debit,
-                Layer::Settled,
-                "USD",
-                account_id,
-            );
-
-            let mut current_balances = HashMap::new();
-            current_balances.insert((account_id, currency), None);
-
-            let entries = vec![entry];
-
-            let result =
-                Balances::new_snapshots(Utc::now(), current_balances, &entries, &HashMap::new());
-
-            assert_eq!(result.len(), 1);
-            let snapshot = &result[0];
-            assert_eq!(snapshot.version, 1); // New snapshot starts at version 1
-        }
-
-        #[test]
-        fn new_snapshots_updates_current_balance_with_entry() {
-            let account_id = AccountId::new();
-            let currency: Currency = "USD".parse().unwrap();
-
-            let mut current_balances = HashMap::new();
-            let version = 5;
-            let mut current_balance =
-                create_test_balance_snapshot(account_id, JournalId::new(), currency, version);
-            current_balance.settled.dr_balance = Decimal::from(200);
-            current_balance.settled.cr_balance = Decimal::from(50);
-            current_balances.insert((account_id, currency), Some(current_balance));
-
-            let entry = create_test_entry(
-                Decimal::from(75),
-                DebitOrCredit::Credit,
-                Layer::Settled,
-                "USD",
-                account_id,
-            );
-            let entries = vec![entry];
-
-            let result =
-                Balances::new_snapshots(Utc::now(), current_balances, &entries, &HashMap::new());
-
-            assert_eq!(result.len(), 1);
-            let snapshot = &result[0];
-            assert_eq!(snapshot.version, version + 1);
-        }
-
-        #[test]
-        fn new_snapshots_can_update_from_multiple_entries() {
-            let account_id = AccountId::new();
-            let journal_id = JournalId::new();
-            let currency: Currency = "USD".parse().unwrap();
-
-            let initial_debit = Decimal::from(100);
-            let initial_credit = Decimal::from(25);
-            let mut current_balances = HashMap::new();
-            let version = 3;
-            let mut current_balance =
-                create_test_balance_snapshot(account_id, journal_id, currency, version);
-            current_balance.settled.dr_balance = initial_debit;
-            current_balance.settled.cr_balance = initial_credit;
-            current_balances.insert((account_id, currency), Some(current_balance));
-
-            // First entry will create a latest balance
-            let entry1_debit = Decimal::from(50);
-            let entry1 = create_test_entry(
-                entry1_debit,
-                DebitOrCredit::Debit,
-                Layer::Settled,
-                "USD",
-                account_id,
-            );
-            // Second entry should use the latest balance, not the current
-            let entry2_credit = Decimal::from(30);
-            let entry2 = create_test_entry(
-                entry2_credit,
-                DebitOrCredit::Credit,
-                Layer::Settled,
-                "USD",
-                account_id,
-            );
-            let entries = vec![entry1, entry2];
-
-            let result =
-                Balances::new_snapshots(Utc::now(), current_balances, &entries, &HashMap::new());
-
-            assert_eq!(result.len(), 2);
-
-            assert_eq!(result[0].version, version + 1);
-            assert_eq!(result[0].settled.dr_balance, initial_debit + entry1_debit);
-
-            assert_eq!(result[1].version, version + 2);
-            assert_eq!(result[1].settled.cr_balance, initial_credit + entry2_credit);
-        }
-
-        #[test]
-        fn new_snapshots_skips_update_when_no_balance_value_exists() {
-            let current_balances = HashMap::new();
-
-            let entry = create_test_entry(
-                Decimal::from(100),
-                DebitOrCredit::Debit,
-                Layer::Settled,
-                "USD",
-                AccountId::new(),
-            );
-            let entries = vec![entry];
-
-            let result =
-                Balances::new_snapshots(Utc::now(), current_balances, &entries, &HashMap::new());
-
-            assert!(result.is_empty());
-        }
-
-        #[test]
-        fn new_snapshots_creates_snapshots_for_mapped_account_sets() {
-            let account_id = AccountId::new();
-            let account_set_id = AccountSetId::new();
-            let currency: Currency = "USD".parse().unwrap();
-
-            let entry = create_test_entry(
-                Decimal::from(100),
-                DebitOrCredit::Debit,
-                Layer::Settled,
-                "USD",
-                account_id,
-            );
-
-            let mut current_balances = HashMap::new();
-            current_balances.insert((account_id, currency), None);
-            current_balances.insert((AccountId::from(&account_set_id), currency), None);
-
-            let mut mappings = HashMap::new();
-            mappings.insert(account_id, vec![account_set_id]);
-
-            let entries = vec![entry];
-
-            let result = Balances::new_snapshots(Utc::now(), current_balances, &entries, &mappings);
-
-            assert_eq!(result.len(), 2);
-        }
-    }
-
-    mod ec_set_snapshots {
-        use super::*;
-
-        use chrono::Utc;
-        use rust_decimal::Decimal;
-        use std::collections::HashMap;
-
-        use cala_types::{
-            entry::EntryValues,
-            primitives::{DebitOrCredit, Layer},
-        };
-
-        use crate::primitives::{AccountSetId, Currency, EntryId, JournalId, TransactionId};
-
-        fn credit_entry(units: Decimal, account_id: AccountId) -> EntryValues {
-            EntryValues {
-                id: EntryId::new(),
-                version: 1,
-                transaction_id: TransactionId::new(),
-                journal_id: JournalId::new(),
-                account_id,
-                entry_type: "TEST_ENTRY".to_string(),
-                sequence: 1,
-                layer: Layer::Settled,
-                currency: "USD".parse().unwrap(),
-                direction: DebitOrCredit::Credit,
-                units,
-                description: None,
-                metadata: None,
-            }
-        }
-
-        #[test]
-        fn fans_member_deltas_into_ec_ancestor_only() {
-            let usd: Currency = "USD".parse().unwrap();
-            let set_id = AccountSetId::new();
-            let set_account = AccountId::from(&set_id);
-            let m1 = AccountId::new();
-            let m2 = AccountId::new();
-
-            let entries = vec![
-                credit_entry(Decimal::from(100), m1),
-                credit_entry(Decimal::from(50), m2),
-            ];
-
-            let mut ec_mappings: HashMap<AccountId, Vec<AccountSetId>> = HashMap::new();
-            ec_mappings.insert(m1, vec![set_id]);
-            ec_mappings.insert(m2, vec![set_id]);
-
-            // EC set has no prior balance.
-            let mut current: HashMap<(AccountId, Currency), Option<BalanceSnapshot>> =
-                HashMap::new();
-            current.insert((set_account, usd), None);
-
-            let snapshots = Balances::ec_set_snapshots(Utc::now(), current, &entries, &ec_mappings);
-
-            // Only the EC set is written — never the leaf accounts.
-            assert!(snapshots.iter().all(|s| s.account_id == set_account));
-            // The final (highest-version) snapshot reflects both credits.
-            let final_snapshot = snapshots.iter().max_by_key(|s| s.version).unwrap();
-            assert_eq!(final_snapshot.settled.cr_balance, Decimal::from(150));
-            assert_eq!(final_snapshot.version, 2);
-        }
-
-        #[test]
-        fn skips_members_without_ec_ancestors() {
-            let entries = vec![credit_entry(Decimal::from(10), AccountId::new())];
-            let ec_mappings: HashMap<AccountId, Vec<AccountSetId>> = HashMap::new();
-            let current: HashMap<(AccountId, Currency), Option<BalanceSnapshot>> = HashMap::new();
-
-            let snapshots = Balances::ec_set_snapshots(Utc::now(), current, &entries, &ec_mappings);
-            assert!(snapshots.is_empty());
-        }
     }
 }
